modular refactor p1
