+++ lib/cli/cli.js
@@ -1,6 +1,9 @@
 import yargs from 'yargs';
 import fs from 'fs';
-const { version } = JSON.parse(fs.readFileSync('../package.json', 'utf8'));
+
+/** @typedef {import('../../types').Dungeon} Config */
+
+const { version } = JSON.parse(fs.readFileSync('../../package.json', 'utf8'));
 
 const hero = String.raw`
