+++ lib/data/defaults.js
@@ -1,8 +1,17 @@
 /* cSpell:disable */
+/**
+ * Default data sets for generating realistic test data
+ * @fileoverview Contains default values for campaigns, devices, locations, and domains
+ */
+
+/** @typedef {import('../../types.d.ts').main.Dungeon} Config */
+/** @typedef {import('../../types.d.ts').main.ValueValid} ValueValid */
+
 //? https://docs.mixpanel.com/docs/data-structure/property-reference#default-properties
 
 const domainSuffix = ["com", "com", "com", "com", "net", "org", "net", "org", "io", "co", "co.uk", "us", "biz", "info", "gov", "edu"];
 const domainPrefix = ["gmail", "gmail", "gmail", "gmail", "gmail", "gmail", "yahoo", "yahoo", "icloud", "icloud", "icloud", "icloud", "hotmail", "hotmail", "gmail", "gmail", "gmail", "gmail", "gmail", "gmail", "yahoo", "yahoo", "icloud", "icloud", "icloud", "hotmail", "hotmail", "outlook", "aol", "outlook", "aol", "protonmail", "zoho", "gmx", "yandex", "mail", "inbox", "fastmail", "tutanota", "mailfence", "disroot", "riseup", "posteo", "runbox", "kolabnow", "mailbox", "scryptmail", "ctemplar", "countermail", "hushmail", "startmail", "privatemail"];
+
 const campaigns = [
 	{
 		utm_campaign: ["$organic"],
@@ -25,7 +34,6 @@ const campaigns = [
 		utm_content: ["sc_control_group", "sc_variant_A", "sc_variant_B", "sc_variant_C", "sc_variant_D"],
 		utm_term: ["sc_jan_feb", "sc_mar_apr", "sc_may_jun", "sc_jul_aug", "sc_sep_oct", "sc_nov_dec"]
 	},
-	
 	{
 		utm_source: ["linkedin"],
 		utm_campaign: ["li_free_trial", "li_discount_US", "li_fall_sale", "li_holiday_special", "li_lookalike_audience"],
@@ -315,7 +323,6 @@ const androidDevices = [
 	}
 ];
 
-
 const desktopDevices = [
 	{
 		model: 'iMac 24-inch (M1, 2021)',
@@ -445,7 +452,6 @@ const desktopDevices = [
 		Platform: 'Desktop'
 	}
 ];
-
 
 const locations = [
 	{
@@ -1241,10 +1247,6 @@ const locations = [
 		"city": "Alexandria"
 	}
 ];
-
-
-
-
 
 export {
 	campaigns,
