+++ lib/utils/ai.js
