+++ lib/utils/utils.js
@@ -10,15 +10,15 @@ import { mkdir, parseGCSUri } from 'ak-tools';
 import { existsSync } from 'fs';
 dayjs.extend(utc);
 import 'dotenv/config';
-import { domainSuffix, domainPrefix } from './defaults.js';
+import { domainSuffix, domainPrefix } from '../data/defaults.js';
 
-/** @typedef {import('../types').Dungeon} Config */
-/** @typedef {import('../types').EventConfig} EventConfig */
-/** @typedef {import('../types').ValueValid} ValueValid */
-/** @typedef {import('../types').HookedArray} hookArray */
-/** @typedef {import('../types').hookArrayOptions} hookArrayOptions */
-/** @typedef {import('../types').Person} Person */
-/** @typedef {import('../types').Funnel} Funnel */
+/** @typedef {import('../../types').Dungeon} Config */
+/** @typedef {import('../../types').EventConfig} EventConfig */
+/** @typedef {import('../../types').ValueValid} ValueValid */
+/** @typedef {import('../../types').HookedArray} hookArray */
+/** @typedef {import('../../types').hookArrayOptions} hookArrayOptions */
+/** @typedef {import('../../types').Person} Person */
+/** @typedef {import('../../types').Funnel} Funnel */
 
 let globalChance;
 let chanceInitialized = false;
