+++ scripts/new-dungeon.sh
@@ -11,7 +11,7 @@ initial_text='
 import dayjs from "dayjs";
 import utc from "dayjs/plugin/utc.js";
 import "dotenv/config";
-import * as u from "../components/utils.js";
+import * as u from "../lib/utils/utils.js";
 import * as v from "ak-tools";
 
 const SEED = "my-seed";
